sd3.5/requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu101
transformers
torch
torchvision
numpy
fire
pillow
numpy
einops
sentencepiece
protobuf
webdataset
